+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Add emulator support to firebase storage MCP tools. (#8707)
 - Removed overeager error catch in `functions:list`.
 - Add validation during `firebase init` feature selection. (#5232)
+- Fixed an issue where the extensions emulator did not work with `demo-` projects. (#8720)
 - Fixed issue where `--export-on-exit` fails if the target directory does not exist. (#4688)